requirements.txt
groq
langchain
langchain-groq
chromadb
sentence-transformers
pymupdf
pdfplumber
PyPDF2
beautifulsoup4
requests
python-dotenv
streamlit
fastapi
uvicorn[standard]
httpx
numpy
pandas
tqdm
pytest
loguru